fix: cast EVE SSO character_id to int (#412)
The eveonline Socialite provider returns attributes.character_id as a string, but
EveUser (declare(strict_types=1)) requires int — login threw a TypeError in
CallbackController. Cast at the boundary. The unit test mocked an int (so it never
caught this); mock the realistic string so it guards the regression.

+++ src/Http/Controllers/Auth/CallbackController.php
@@ -29,7 +29,8 @@ class CallbackController
         $socialite_user = $social->driver('eveonline')->user();
 
         $eve_data = new EveUser(
-            character_id: data_get($socialite_user, 'attributes.character_id'),
+            // EVE SSO returns character_id as a string; EveUser (strict_types) expects int.
+            character_id: (int) data_get($socialite_user, 'attributes.character_id'),
             character_owner_hash: data_get($socialite_user, 'attributes.character_owner_hash'),
             token: data_get($socialite_user, 'token'),
             refreshToken: data_get($socialite_user, 'refreshToken'),
